Original from Ross Scrivener
This is a Javascript implementation of the Langton ant program.  He
allows reuse, "This example is made using Javascript and the Canvas API,
pretty simple stuff, you can view source to steal the code should you
wish, it's slightly over complicated because I wanted to add the control
panel, a lot of it can be stripped down."

This code should be a base for my modified version.